# Refactor geometric_soft_configuration_model tests for performance