#6858 added the `geometric_soft_configuration_model` generator, which included several tests that generated graphs with >1k nodes. These tests were quite slow:

```bash
$ pytest --durations 4 networkx/generators/tests/test_geometric.py

...

4.07s call     networkx/generators/tests/test_geometric.py::test_mean_degree_S1
4.06s call     networkx/generators/tests/test_geometric.py::test_mean_kappas_S1
1.34s call     networkx/generators/tests/test_geometric.py::test_compare_mean_kappas_different_gammas_S1
0.18s call     networkx/generators/tests/test_geometric.py::test_dict_kappas_S1
```

These could be decorated with `pytest.mark.slow`, but then they'd still be running in the coverage job in CI. Since the tests are pseudo-stochastic anyways (i.e. they test average properties of graphs created with a seed) I figured there was no harm in refactoring them instead to improve performance. The main thing was to drop the total number of nodes from 1000's to 10's (reducing the `mean_degree` as well, where necessary). I also consolidated two of the tests into one so that the input graph is only created once. These changes result in:

```bash
$ pytest --durations 4 networkx/generators/tests/test_geometric.py

...

0.18s call     networkx/generators/tests/test_geometric.py::test_dict_kappas_S1
0.03s call     networkx/generators/tests/test_geometric.py::TestRandomGeometricGraph::test_number_of_nodes
0.03s call     networkx/generators/tests/test_geometric.py::TestNavigableSmallWorldGraph::test_navigable_small_world
0.01s call     networkx/generators/tests/test_geometric.py::test_beta_clustering_S1
```

... removing the bottlenecks and dropping the total test runtime for the module from ~10s -> 0.4s (on my machine).