@rossbar 2024-01-05
In CI the savings are even more pronounced: the total test runtime on ubuntu is about 30 sec shorter (25%) [for this branch](https://github.com/networkx/networkx/actions/runs/7426652814/job/20210827992?pr=7210) compared to [main](https://github.com/networkx/networkx/actions/runs/7413651948/job/20173063306).